# swagger/swagger20.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "gopaytest code test payments API",
    "version" : "1.0.0",
    "title" : "Payments API",
    "termsOfService" : "",
    "contact" : {
      "name" : "Francesco Casula",
      "email" : "fra.casula@gmail.com"
    }
  },
  "host" : "api.test.gopaytest.tech",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/payments/" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Get the list of Payments",
        "description" : "Loads a list of Payments",
        "parameters" : [ {
          "name" : "$size",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "description" : "Size of the page to retrieve.",
          "x-example" : 10
        }, {
          "name" : "$page",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "description" : "Number of the page to retrieve.",
          "x-example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Payment"
              }
            },
            "examples" : {
              "application/json" : "{\n    \"data\": [\n        {\n            \"type\": \"Payment\",\n            \"id\": \"4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43\",\n            \"version\": 0,\n            \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n            \"attributes\": {\n                \"amount\": \"100.21\",\n                \"beneficiary_party\": {\n                    \"account_name\": \"W Owens\",\n                    \"account_number\": \"31926819\",\n                    \"account_number_code\": \"BBAN\",\n                    \"account_type\": 0,\n                    \"address\": \"1 The Beneficiary Localtown SE2\",\n                    \"bank_id\": \"403000\",\n                    \"bank_id_code\": \"GBDSC\",\n                    \"name\": \"Wilfred Jeremiah Owens\"\n                },\n                \"charges_information\": {\n                    \"bearer_code\": \"SHAR\",\n                    \"sender_charges\": [\n                        {\n                            \"amount\": \"5.00\",\n                            \"currency\": \"GBP\"\n                        },\n                        {\n                            \"amount\": \"10.00\",\n                            \"currency\": \"USD\"\n                        }\n                    ],\n                    \"receiver_charges_amount\": \"1.00\",\n                    \"receiver_charges_currency\": \"USD\"\n                },\n                \"currency\": \"GBP\",\n                \"debtor_party\": {\n                    \"account_name\": \"EJ Brown Black\",\n                    \"account_number\": \"GB29XABC10161234567801\",\n                    \"account_number_code\": \"IBAN\",\n                    \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                    \"bank_id\": \"203301\",\n                    \"bank_id_code\": \"GBDSC\",\n                    \"name\": \"Emelia Jane Brown\"\n                },\n                \"end_to_end_reference\": \"Wil piano Jan\",\n                \"fx\": {\n                    \"contract_reference\": \"FX123\",\n                    \"exchange_rate\": \"2.00000\",\n                    \"original_amount\": \"200.42\",\n                    \"original_currency\": \"USD\"\n                },\n                \"numeric_reference\": \"1002001\",\n                \"payment_id\": \"123456789012345678\",\n                \"payment_purpose\": \"Paying for goods/services\",\n                \"payment_scheme\": \"FPS\",\n                \"payment_type\": \"Credit\",\n                \"processing_date\": \"2017-01-18\",\n                \"reference\": \"Payment for Em's piano lessons\",\n                \"scheme_payment_sub_type\": \"InternetBanking\",\n                \"scheme_payment_type\": \"ImmediatePayment\",\n                \"sponsor_party\": {\n                    \"account_number\": \"56781234\",\n                    \"bank_id\": \"123123\",\n                    \"bank_id_code\": \"GBDSC\"\n                }\n            }\n        }\n    ],\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments\"\n    }\n}\n"
            },
            "headers" : {
              "X-Page-Count" : {
                "type" : "integer",
                "x-example" : 1
              },
              "X-Page-Number" : {
                "type" : "integer",
                "x-example" : 1
              },
              "X-Page-Size" : {
                "type" : "integer",
                "x-example" : 25
              },
              "X-Total-Count" : {
                "type" : "integer",
                "x-example" : 2
              }
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Create a Payment",
        "description" : "Adds a Payment",
        "consumes" : [ ],
        "parameters" : [ {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Payment"
          },
          "x-examples" : {
            "application/json" : "{\n    \"type\": \"Payment\",\n    \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n    \"attributes\": {\n        \"amount\": \"100.21\",\n        \"beneficiary_party\": {\n            \"account_name\": \"W Owens\",\n            \"account_number\": \"31926819\",\n            \"account_number_code\": \"BBAN\",\n            \"account_type\": 0,\n            \"address\": \"1 The Beneficiary Localtown SE2\",\n            \"bank_id\": \"403000\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Wilfred Jeremiah Owens\"\n        },\n        \"charges_information\": {\n            \"bearer_code\": \"SHAR\",\n            \"sender_charges\": [\n                {\n                    \"amount\": \"5.00\",\n                    \"currency\": \"GBP\"\n                },\n                {\n                    \"amount\": \"10.00\",\n                    \"currency\": \"USD\"\n                }\n            ],\n            \"receiver_charges_amount\": \"1.00\",\n            \"receiver_charges_currency\": \"USD\"\n        },\n        \"currency\": \"GBP\",\n        \"debtor_party\": {\n            \"account_name\": \"EJ Brown Black\",\n            \"account_number\": \"GB29XABC10161234567801\",\n            \"account_number_code\": \"IBAN\",\n            \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n            \"bank_id\": \"203301\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Emelia Jane Brown\"\n        },\n        \"end_to_end_reference\": \"Wil piano Jan\",\n        \"fx\": {\n            \"contract_reference\": \"FX123\",\n            \"exchange_rate\": \"2.00000\",\n            \"original_amount\": \"200.42\",\n            \"original_currency\": \"USD\"\n        },\n        \"numeric_reference\": \"1002001\",\n        \"payment_id\": \"123456789012345678\",\n        \"payment_purpose\": \"Paying for goods/services\",\n        \"payment_scheme\": \"FPS\",\n        \"payment_type\": \"Credit\",\n        \"processing_date\": \"2017-01-18\",\n        \"reference\": \"Payment for Em's piano lessons\",\n        \"scheme_payment_sub_type\": \"InternetBanking\",\n        \"scheme_payment_type\": \"ImmediatePayment\",\n        \"sponsor_party\": {\n            \"account_number\": \"56781234\",\n            \"bank_id\": \"123123\",\n            \"bank_id_code\": \"GBDSC\"\n        }\n    }\n}"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Status 201",
            "schema" : {
              "$ref" : "#/definitions/Payment"
            },
            "examples" : {
              "application/json" : "{\n    \"data\": {\n        \"type\": \"Payment\",\n        \"id\": \"cca3b50b-4acb-43dd-8465-8e7d6eac2cac\",\n        \"version\": 1,\n        \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n        \"attributes\": {\n            \"amount\": \"100.21\",\n            \"beneficiary_party\": {\n                \"account_name\": \"W Owens\",\n                \"account_number\": \"31926819\",\n                \"account_number_code\": \"BBAN\",\n                \"account_type\": 0,\n                \"address\": \"1 The Beneficiary Localtown SE2\",\n                \"bank_id\": \"403000\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Wilfred Jeremiah Owens\"\n            },\n            \"charges_information\": {\n                \"bearer_code\": \"SHAR\",\n                \"sender_charges\": [\n                    {\n                        \"amount\": \"5.00\",\n                        \"currency\": \"GBP\"\n                    },\n                    {\n                        \"amount\": \"10.00\",\n                        \"currency\": \"USD\"\n                    }\n                ],\n                \"receiver_charges_amount\": \"1.00\",\n                \"receiver_charges_currency\": \"USD\"\n            },\n            \"currency\": \"GBP\",\n            \"debtor_party\": {\n                \"account_name\": \"EJ Brown Black\",\n                \"account_number\": \"GB29XABC10161234567801\",\n                \"account_number_code\": \"IBAN\",\n                \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                \"bank_id\": \"203301\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Emelia Jane Brown\"\n            },\n            \"end_to_end_reference\": \"Wil piano Jan\",\n            \"fx\": {\n                \"contract_reference\": \"FX123\",\n                \"exchange_rate\": \"2.00000\",\n                \"original_amount\": \"200.42\",\n                \"original_currency\": \"USD\"\n            },\n            \"numeric_reference\": \"1002001\",\n            \"payment_id\": \"123456789012345678\",\n            \"payment_purpose\": \"Paying for goods/services\",\n            \"payment_scheme\": \"FPS\",\n            \"payment_type\": \"Credit\",\n            \"processing_date\": \"2017-01-18\",\n            \"reference\": \"Payment for Em's piano lessons\",\n            \"scheme_payment_sub_type\": \"InternetBanking\",\n            \"scheme_payment_type\": \"ImmediatePayment\",\n            \"sponsor_party\": {\n                \"account_number\": \"56781234\",\n                \"bank_id\": \"123123\",\n                \"bank_id_code\": \"GBDSC\"\n            }\n        }\n    },\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments/cca3b50b-4acb-43dd-8465-8e7d6eac2cac\"\n    }\n}"
            },
            "headers" : {
              "Location" : {
                "type" : "string",
                "x-example" : "https://api.test.gopaytest.tech/v1/payments/4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43"
              }
            }
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "None" : {
      "description" : "This API is not secured, it's not meant to be exposed outside of the cluster. Authentication should be handled via another API (e.g. API Gateway)."
    }
  },
  "definitions" : {
    "Error" : {
      "type" : "object",
      "required" : [ "code", "reasonPhrase" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "minimum" : 400,
          "maximum" : 599
        },
        "description" : {
          "type" : "string",
          "example" : "Bad query parameter [$size]: Invalid integer value [abc]"
        },
        "reasonPhrase" : {
          "type" : "string",
          "example" : "Bad Request"
        }
      },
      "description" : "This general error structure is used throughout this API.",
      "example" : "{\n  \"code\": 400,\n  \"description\": \"Bad query parameter [$size]: Invalid integer value [abc]\",\n  \"reasonPhrase\": \"Bad Request\"\n}"
    },
    "Payment" : {
      "type" : "object",
      "required" : [ "attributes", "id", "organisation_id", "type", "version" ],
      "properties" : {
        "type" : {
          "type" : "string",
          "enum" : [ "Payment" ]
        },
        "id" : {
          "type" : "string",
          "description" : "Auto-generated primary key field",
          "example" : "0e8ffb10-ad98-11e6-bf2e-47644ada7c0f"
        },
        "version" : {
          "type" : "number",
          "default" : 0.0
        },
        "organisation_id" : {
          "type" : "string",
          "example" : "743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb"
        },
        "attributes" : {
          "required" : [ "amount", "beneficiary_party", "charges_information", "currency", "debtor_party", "end_to_end_reference", "fx", "numeric_reference", "payment_id", "payment_purpose", "payment_scheme", "payment_type", "processing_date", "reference", "scheme_payment_sub_type", "scheme_payment_type" ],
          "type" : "object",
          "properties" : {
            "amount" : {
              "type" : "string",
              "example" : "100.21"
            },
            "beneficiary_party" : {
              "required" : [ "account_name", "account_number", "account_number_code", "account_type", "address", "bank_id", "bank_id_code", "name" ],
              "type" : "object",
              "properties" : {
                "account_name" : {
                  "type" : "string"
                },
                "account_number" : {
                  "type" : "string"
                },
                "account_number_code" : {
                  "type" : "string"
                },
                "account_type" : {
                  "type" : "integer",
                  "default" : 0
                },
                "address" : {
                  "type" : "string"
                },
                "bank_id" : {
                  "type" : "string"
                },
                "bank_id_code" : {
                  "type" : "string"
                },
                "name" : {
                  "type" : "string"
                }
              }
            },
            "charges_information" : {
              "required" : [ "bearer_code", "receiver_charges_amount", "receiver_charges_currency", "sender_charges" ],
              "type" : "object",
              "properties" : {
                "bearer_code" : {
                  "type" : "string",
                  "example" : "SHAR"
                },
                "sender_charges" : {
                  "type" : "array",
                  "items" : {
                    "required" : [ "amount", "currency" ],
                    "type" : "object",
                    "properties" : {
                      "amount" : {
                        "type" : "string",
                        "default" : "5.00"
                      },
                      "currency" : {
                        "type" : "string",
                        "example" : "GBP"
                      }
                    }
                  }
                },
                "receiver_charges_amount" : {
                  "type" : "string",
                  "default" : "1.00"
                },
                "receiver_charges_currency" : {
                  "type" : "string",
                  "example" : "GBP"
                }
              }
            },
            "currency" : {
              "type" : "string",
              "example" : "GBP"
            },
            "debtor_party" : {
              "required" : [ "account_name", "account_number", "account_number_code", "address", "bank_id", "bank_id_code", "name" ],
              "type" : "object",
              "properties" : {
                "account_name" : {
                  "type" : "string"
                },
                "account_number" : {
                  "type" : "string"
                },
                "account_number_code" : {
                  "type" : "string"
                },
                "address" : {
                  "type" : "string"
                },
                "bank_id" : {
                  "type" : "string"
                },
                "bank_id_code" : {
                  "type" : "string"
                },
                "name" : {
                  "type" : "string"
                }
              }
            },
            "end_to_end_reference" : {
              "type" : "string"
            },
            "fx" : {
              "required" : [ "contract_reference", "exchange_rate", "original_amount", "original_currency" ],
              "type" : "object",
              "properties" : {
                "contract_reference" : {
                  "type" : "string"
                },
                "exchange_rate" : {
                  "type" : "string",
                  "example" : "2.00000"
                },
                "original_amount" : {
                  "type" : "string",
                  "example" : "200.42"
                },
                "original_currency" : {
                  "type" : "string",
                  "example" : "GBP"
                }
              }
            },
            "numeric_reference" : {
              "type" : "string"
            },
            "payment_id" : {
              "type" : "string"
            },
            "payment_purpose" : {
              "type" : "string"
            },
            "payment_scheme" : {
              "type" : "string"
            },
            "payment_type" : {
              "type" : "string"
            },
            "processing_date" : {
              "type" : "string",
              "format" : "date",
              "example" : "2017-01-18"
            },
            "reference" : {
              "type" : "string"
            },
            "scheme_payment_sub_type" : {
              "type" : "string",
              "example" : "InternetBanking"
            },
            "scheme_payment_type" : {
              "type" : "string",
              "example" : "ImmediatePayment"
            },
            "sponsor_party" : {
              "required" : [ "account_number", "bank_id", "bank_id_code" ],
              "type" : "object",
              "properties" : {
                "account_number" : {
                  "type" : "string"
                },
                "bank_id" : {
                  "type" : "string"
                },
                "bank_id_code" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      },
      "description" : "A representation of a Payment.",
      "example" : "{\"type\":\"Payment\",\"id\":\"4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43\",\"version\":0,\"organisation_id\":\"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\"attributes\":{\"amount\":\"100.21\",\"beneficiary_party\":{\"account_name\":\"W Owens\",\"account_number\":\"31926819\",\"account_number_code\":\"BBAN\",\"account_type\":0,\"address\":\"1 The Beneficiary Localtown SE2\",\"bank_id\":\"403000\",\"bank_id_code\":\"GBDSC\",\"name\":\"Wilfred Jeremiah Owens\"},\"charges_information\":{\"bearer_code\":\"SHAR\",\"sender_charges\":[{\"amount\":\"5.00\",\"currency\":\"GBP\"},{\"amount\":\"10.00\",\"currency\":\"USD\"}],\"receiver_charges_amount\":\"1.00\",\"receiver_charges_currency\":\"USD\"},\"currency\":\"GBP\",\"debtor_party\":{\"account_name\":\"EJ Brown Black\",\"account_number\":\"GB29XABC10161234567801\",\"account_number_code\":\"IBAN\",\"address\":\"10 Debtor Crescent Sourcetown NE1\",\"bank_id\":\"203301\",\"bank_id_code\":\"GBDSC\",\"name\":\"Emelia Jane Brown\"},\"end_to_end_reference\":\"Wil piano Jan\",\"fx\":{\"contract_reference\":\"FX123\",\"exchange_rate\":\"2.00000\",\"original_amount\":\"200.42\",\"original_currency\":\"USD\"},\"numeric_reference\":\"1002001\",\"payment_id\":\"123456789012345678\",\"payment_purpose\":\"Paying for goods/services\",\"payment_scheme\":\"FPS\",\"payment_type\":\"Credit\",\"processing_date\":\"2017-01-18\",\"reference\":\"Payment for Em's piano lessons\",\"scheme_payment_sub_type\":\"InternetBanking\",\"scheme_payment_type\":\"ImmediatePayment\",\"sponsor_party\":{\"account_number\":\"56781234\",\"bank_id\":\"123123\",\"bank_id_code\":\"GBDSC\"}}}"
    }
  }
}